requirements.txt
fastapi
uvicorn
python-multipart
pillow
PyMuPDF==1.26.3
